Move icon to its own files to fix the build
